refactor(playground): split types/ module
Move types.ts into src/types/ with four focused files:
- sim-dto.ts: Car, Stop, Snapshot, Metrics
- strategies.ts: StrategyName, RepositionStrategyName, TrafficMode
- bubble.ts: BubbleEvent, CarBubble
- scenario.ts: Phase, ElevatorPhysics, TweakRange(s), ScenarioMeta

Barrel re-exports all types so existing `from "./types"` imports
resolve unchanged via bundler module resolution.
