+++ playground/src/types/index.ts
@@ -1,0 +1,4 @@
+export type { Car, Stop, Snapshot, Metrics } from "./sim-dto";
+export type { StrategyName, RepositionStrategyName, TrafficMode } from "./strategies";
+export type { BubbleEvent, CarBubble } from "./bubble";
+export type { Phase, ElevatorPhysics, TweakRange, TweakRanges, ScenarioMeta } from "./scenario";
