+++ playground/src/types/scenario.ts
@@ -1,130 +1,4 @@
-// Mirrors of the DTO shapes returned by crates/elevator-wasm/src/dto.rs.
-// Keep this file narrow — only what the UI actually reads.
-
-export interface Car {
-  id: number;
-  line: number;
-  y: number;
-  v: number;
-  phase:
-    | "idle"
-    | "moving"
-    | "repositioning"
-    | "door-opening"
-    | "loading"
-    | "door-closing"
-    | "stopped"
-    | "unknown";
-  target: number | null;
-  load: number;
-  capacity: number;
-  riders: number;
-  /**
-   * Min and max world-y of stops the car's line serves. Renderers
-   * use these to draw a shaft channel that only spans the range the
-   * car can reach — express banks and service elevators get visibly
-   * shorter shafts than the full-building banks. `NaN` when the
-   * wasm couldn't resolve a range (stale build or mis-configured
-   * line); renderers fall back to the full canvas height in that case.
-   */
-  min_served_y: number;
-  max_served_y: number;
-}
-
-export interface Stop {
-  entity_id: number;
-  stop_id: number;
-  name: string;
-  y: number;
-  waiting: number;
-  /** Waiting riders whose destination is above this stop. Partition of `waiting`. */
-  waiting_up: number;
-  /** Waiting riders whose destination is below this stop. Partition of `waiting`. */
-  waiting_down: number;
-  residents: number;
-}
-
-export interface Snapshot {
-  tick: number;
-  dt: number;
-  cars: Car[];
-  stops: Stop[];
-}
-
-export interface Metrics {
-  delivered: number;
-  abandoned: number;
-  spawned: number;
-  settled: number;
-  rerouted: number;
-  throughput: number;
-  avg_wait_s: number;
-  max_wait_s: number;
-  avg_ride_s: number;
-  utilization: number;
-  abandonment_rate: number;
-  total_distance: number;
-  total_moves: number;
-}
-
-export type StrategyName = "scan" | "look" | "nearest" | "etd" | "destination" | "rsr";
-
-/**
- * Per-pane reposition (idle-parking) strategy. Mirrors the five
- * built-ins exposed by the wasm crate's `builtinRepositionStrategies`;
- * kept as a string literal union so typos surface at the compiler.
- *
- * - `adaptive` — mode-gated: ReturnToLobby in up-peak, PredictiveParking
- *   otherwise. The playground's default.
- * - `predictive` — always park near the hottest recent-arrival stop.
- * - `lobby` — always return to stop 0.
- * - `spread` — maximise inter-car spacing across the shaft.
- * - `none` — stay where the car stopped.
- */
-export type RepositionStrategyName = "adaptive" | "predictive" | "lobby" | "spread" | "none";
-
-/**
- * Traffic-mode readout from the core `TrafficDetector`, surfaced by
- * `Sim.trafficMode()`. Mirrors the Rust `TrafficMode` enum 1:1.
- * `AdaptiveParking` reads this each reposition pass to pick between
- * `ReturnToLobby` / `PredictiveParking` / no-op — exposing it in the UI
- * lets users see *why* idle-car movement changes over the course of a
- * scenario.
- */
-export type TrafficMode = "Idle" | "UpPeak" | "InterFloor" | "DownPeak";
-
-/**
- * Decoded event DTOs surfaced by `Sim.drainEvents`. Kind-tagged to
- * mirror the Rust `EventDto` shape (`#[serde(tag = "kind", rename_all =
- * "kebab-case")]`). Only the cases the speech-bubble layer consumes are
- * enumerated; unknown variants fall through to the `string`-kind branch
- * so a future DTO addition doesn't crash the UI.
- */
-export type BubbleEvent =
-  | { kind: "rider-spawned"; tick: number; rider: number; origin: number; destination: number }
-  | { kind: "rider-boarded"; tick: number; rider: number; elevator: number }
-  | { kind: "rider-exited"; tick: number; rider: number; elevator: number; stop: number }
-  | { kind: "rider-abandoned"; tick: number; rider: number; stop: number }
-  | { kind: "elevator-arrived"; tick: number; elevator: number; stop: number }
-  | { kind: "elevator-departed"; tick: number; elevator: number; stop: number }
-  | { kind: "door-opened"; tick: number; elevator: number }
-  | { kind: "door-closed"; tick: number; elevator: number }
-  | { kind: "elevator-assigned"; tick: number; elevator: number; stop: number }
-  | { kind: "elevator-repositioning"; tick: number; elevator: number; stop: number }
-  | { kind: "other"; tick: number; label: string };
-
-/**
- * Per-car speech-bubble state, keyed by car entity id. All three
- * timestamps use `performance.now()` wall-clock ms — not sim ticks — so
- * the bubble fades predictably even when the sim races ahead or the tab
- * backgrounds and `requestAnimationFrame` stalls. `bornAt` exists so
- * the renderer can fade the last ~30 % of lifetime for a soft exit.
- */
-export interface CarBubble {
-  text: string;
-  bornAt: number;
-  expiresAt: number;
-}
+import type { StrategyName } from "./strategies";
 
 /**
  * One phase of a scenario's day cycle. The `TrafficDriver` linearly
